Api services tested, Accessible-location-services created, project refactored
