+++ mobile/src/app/accessible-location/constants/accessible-location.constants.ts
@@ -1,0 +1,8 @@
+/**
+ * Created by marcosflavio on 6/18/17.
+ */
+export const ACCESSIBLELOCATION = {
+    ACCESSIBILITYTYPE: "/acessibilityType",
+    COMMENT: "/comment",
+    MARKER: "/marker"
+};
