+++ mobile/src/app/commons/api/constants/server.constants.ts
