+++ mobile/src/app/commons/google/constants/google-geolocation.constants.ts
@@ -1,0 +1,6 @@
+/**
+ * Created by marcosflavio on 6/18/17.
+ */
+export const GEOLOCATION = {
+    DOGEOCODE: "/geolocation"
+};
