+++ mobile/src/app/commons/google/services/google-geolocation.service.ts
@@ -2,19 +2,20 @@
  * Created by marcosflavio on 6/15/17.
  */
 import {Injectable} from "@angular/core";
-import {HttpService} from "../../api/http.service";
+import {HttpService} from "../../api/services/http.service";
 import {Observable} from "rxjs";
+import {GEOLOCATION} from "../constants/google-geolocation.constants";
 
 @Injectable()
 export class GoogleGeoLocationService {
 
-    private geoLococationServiceUrl = "/geolocation";
+    private doGeocodeUrl = GEOLOCATION.DOGEOCODE;
 
     constructor(private httpService: HttpService) {
     }
 
     doGeocode(address: string): Observable<any> {
-        return this.httpService.post(this.geoLococationServiceUrl, address)
+        return this.httpService.post(this.doGeocodeUrl, address)
             .map(result => result.json());
     }
 }
